improve underlying lexer performance
