--- lexer/pool.go
@@ -1,12 +1,0 @@
-package lexer
-
-import (
-	"bytes"
-	"sync"
-)
-
-var pool = sync.Pool{
-	New: func() any {
-		return new(bytes.Buffer)
-	},
-}
